fuck

+++ airpump_test_v01/airpump.cpp
@@ -1,7 +1,7 @@
 #ifndef _SOURCE_
 #define _SOURCE_
 
-int pressure_check(int val, int high, int low) {
+int pressure_check(float val, float high, float low) {
   if (low > val) return 1;
   if (high < val) return -1;
   return 0;
